BACKEND/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langgraph
langchain
langchain-huggingface
supabase
python-dotenv
httpx
python-multipart
